Created forcevectorcalculator class
