+++ implementation_and_examples/agent_implementation/agent.cpp
@@ -411,480 +411,7 @@ void Agent::checkIfAgentFitsBetweenObstacles(quadtree::Box objectBox) const {
 
 
 
-/**
- * Calculate the vector to avoid objects:
- * Find the closest relative angle that is free of objects within a certain range.
- * According to "Dynamic Frontier-Led Swarming: Multi-Robot Repeated Coverage in Dynamic Environments" paper
- * https://ieeexplore-ieee-org.tudelft.idm.oclc.org/stamp/stamp.jsp?tp=&arnumber=10057179&tag=1
- * @return The vector towards the free direction
- */
-bool Agent::calculateObjectAvoidanceAngle(argos::CRadians *relativeObjectAvoidanceAngle, argos::CRadians targetAngle) {
 
-    //Get occupied boxes within range
-    std::vector<quadtree::Box> occupiedBoxes = this->quadtree->queryOccupiedBoxes(this->position,
-                                                                                  OBJECT_AVOIDANCE_RADIUS * 2.0,
-                                                                                  this->elapsed_ticks /
-                                                                                  this->ticks_per_second);
-
-    double angleInterval = argos::CDegrees(360 / ANGLE_INTERVAL_STEPS).GetValue();
-
-    //Create set of free angles ordered to be used for wall following
-    std::set<argos::CDegrees, CustomComparator> freeAngles(
-            CustomComparator(wallFollower.wallFollowingDirection, ToDegrees(heading).GetValue(), ToDegrees(targetAngle).GetValue()));
-
-//Add free angles from -180 to 180 degrees
-    for (int a = 0; a < ANGLE_INTERVAL_STEPS; a++) {
-        auto angle = argos::CDegrees(a * 360 / ANGLE_INTERVAL_STEPS - 180);
-        freeAngles.insert(angle);
-    }
-
-
-    //For each occupied box, find the angles that are blocked relative to the agent
-    for (auto box: occupiedBoxes) {
-
-
-        argos::CVector2 OC = argos::CVector2(box.getCenter().x - this->position.x,
-                                             box.getCenter().y - this->position.y);
-        argos::CRadians Bq = argos::ASin(
-                std::min(AGENT_SAFETY_RADIUS + OBJECT_SAFETY_RADIUS, OC.Length()) / OC.Length());
-        argos::CRadians Eta_q = OC.Angle();
-        if (AGENT_SAFETY_RADIUS + OBJECT_SAFETY_RADIUS > OC.Length())
-            argos::RLOGERR << "AGENT_SAFETY_RADIUS + OBJECT_SAFETY_RADIOS > OC.Length(): " << AGENT_SAFETY_RADIUS
-                           << " + " << OBJECT_SAFETY_RADIUS << ">" << OC.Length() << std::endl;
-
-        argos::CDegrees minAngle = ToDegrees((Eta_q - Bq).SignedNormalize());
-        argos::CDegrees maxAngle = ToDegrees((Eta_q + Bq).SignedNormalize());
-
-        if (maxAngle < minAngle) {
-            argos::CDegrees temp = minAngle;
-            minAngle = maxAngle;
-            maxAngle = temp;
-        }
-
-
-        //Round to multiples of 360/ANGLE_INTERVAL_STEPS
-        double minAngleVal = minAngle.GetValue();
-        double intervalDirectionMin = (minAngleVal < 0) ? -angleInterval : angleInterval;
-//
-        auto minRoundedAngle1 = (int) (minAngleVal / angleInterval) * angleInterval;
-        auto minRoundedAngle2 = minRoundedAngle1 + intervalDirectionMin;
-
-        auto roundedMinAngle = argos::CDegrees(
-                (abs(minAngleVal - minRoundedAngle1) >= abs(minRoundedAngle2 - minAngleVal)) ? minRoundedAngle2
-                                                                                             : minRoundedAngle1);
-
-        double maxAngleVal = maxAngle.GetValue();
-        double intervalDirectionMax = (maxAngleVal < 0) ? -angleInterval : angleInterval;
-
-        auto maxRoundedAngle1 = (int) (maxAngleVal / angleInterval) * angleInterval;
-        auto maxRoundedAngle2 = maxRoundedAngle1 + intervalDirectionMax;
-
-        auto roundedMaxAngle = argos::CDegrees(
-                (abs(maxAngleVal - maxRoundedAngle1) >= abs(maxRoundedAngle2 - maxAngleVal)) ? maxRoundedAngle2
-                                                                                             : maxRoundedAngle1);
-
-        //Block all angles within range
-        for (int a = 0; a < ANGLE_INTERVAL_STEPS; a++) {
-            auto angle = argos::CDegrees(a * 360 / ANGLE_INTERVAL_STEPS - 180);
-
-            if (NormalizedDifference(ToRadians(roundedMinAngle), Eta_q) <= argos::CRadians(0) &&
-                NormalizedDifference(ToRadians(roundedMaxAngle), Eta_q) >= argos::CRadians(0)) {
-                if (angle >= roundedMinAngle && angle <= roundedMaxAngle) {
-                    freeAngles.erase(angle);
-                }
-
-            } else if (NormalizedDifference(ToRadians(roundedMinAngle), Eta_q) >= argos::CRadians(0) &&
-                       NormalizedDifference(ToRadians(roundedMaxAngle), Eta_q) <= argos::CRadians(0)) {
-
-                if (angle <= roundedMinAngle || angle >= roundedMaxAngle) {
-                    freeAngles.erase(angle);
-                }
-            } else {
-                argos::LOGERR << "Error: Eta_q not within range" << std::endl;
-            }
-        }
-
-    }
-
-    this->freeAnglesVisualization.clear();
-    for (auto freeAngle: freeAngles) {
-        this->freeAnglesVisualization.insert(freeAngle);
-    }
-    if (freeAngles.empty()) return false;
-
-
-    //Get free angle closest to heading
-    auto closestFreeAngle = *freeAngles.begin();
-#ifdef WALL_FOLLOWING_ENABLED
-    if (wallFollower.wallFollowingDirection != 0) { //If wall following is not 0, find the closest free angle to the target angle.
-        for (auto freeAngle: freeAngles) {
-            if (std::abs(NormalizedDifference(ToRadians(freeAngle), targetAngle).GetValue()) <
-                std::abs(NormalizedDifference(ToRadians(closestFreeAngle), targetAngle).GetValue())) {
-                closestFreeAngle = freeAngle;
-            }
-        }
-    }
-#endif
-
-
-    argos::CRadians closestFreeAngleRadians = ToRadians(closestFreeAngle);
-    *relativeObjectAvoidanceAngle = NormalizedDifference(closestFreeAngleRadians, targetAngle);
-
-#ifdef WALL_FOLLOWING_ENABLED//If wall following is enabled
-    wallFollower.wallFollowing(this, freeAngles, &closestFreeAngle, &closestFreeAngleRadians, relativeObjectAvoidanceAngle, targetAngle);
-#endif
-
-    return true;
-
-}
-
-#ifdef WALL_FOLLOWING_ENABLED
-
-#endif
-
-/** Calculate the vector to avoid the virtual walls
- * If the agent is close to the border, create a vector pointing away from the border
- * Implemented according to "Dynamic Frontier-Led Swarming: Multi-Robot Repeated Coverage in Dynamic Environments" paper
- * https://ieeexplore-ieee-org.tudelft.idm.oclc.org/stamp/stamp.jsp?tp=&arnumber=10057179&tag=1
- * However, the vector directions are flipped compared to the paper, as the paper uses a different coordinate system
- * @return a vector pointing away from the border
- *
- */
-argos::CVector2 Agent::getVirtualWallAvoidanceVector() const {
-    //If the agent is close to the border, create a vector pointing away from the border
-    argos::CVector2 virtualWallAvoidanceVector = {0, 0};
-
-    if (this->position.x < this->left_right_borders.x) {
-        virtualWallAvoidanceVector.SetX(1);
-    } else if (this->left_right_borders.x <= this->position.x && this->position.x <= this->left_right_borders.y) {
-        virtualWallAvoidanceVector.SetX(0);
-    } else if (this->position.x > this->left_right_borders.y) {
-        virtualWallAvoidanceVector.SetX(-1);
-    }
-
-    if (this->position.y < this->upper_lower_borders.y) {
-        virtualWallAvoidanceVector.SetY(1);
-    } else if (this->upper_lower_borders.y <= this->position.y && this->position.y <= this->upper_lower_borders.x) {
-        virtualWallAvoidanceVector.SetY(0);
-    } else if (this->position.y > this->upper_lower_borders.x) {
-        virtualWallAvoidanceVector.SetY(-1);
-    }
-
-
-    return virtualWallAvoidanceVector;
-
-}
-
-bool Agent::getAverageNeighborLocation(Coordinate *averageNeighborLocation, double range) {
-    int nAgentsWithinRange = 0;
-    for (const auto &agentLocation: this->agentLocations) {
-        if((agentLocation.second.second - this->elapsed_ticks) / this->ticks_per_second > AGENT_LOCATION_RELEVANT_DURATION_S) continue;
-        argos::CVector2 vectorToOtherAgent =
-                argos::CVector2(agentLocation.second.first.x, agentLocation.second.first.y)
-                - argos::CVector2(this->position.x, this->position.y);
-
-        if (vectorToOtherAgent.Length() < range) {
-            averageNeighborLocation->x += agentLocation.second.first.x;
-            averageNeighborLocation->y += agentLocation.second.first.y;
-            nAgentsWithinRange++;
-        }
-    }
-    //If no agents are within range, there is no average location
-    if (nAgentsWithinRange == 0) {
-        return false;
-    }
-    //Else calculate the average position of the agents within range
-    averageNeighborLocation->x /= nAgentsWithinRange;
-    averageNeighborLocation->y /= nAgentsWithinRange;
-
-    return true;
-}
-
-argos::CVector2 Agent::calculateAgentCohesionVector() {
-    Coordinate averageNeighborLocation = {0, 0};
-
-    bool neighborsWithinRange = getAverageNeighborLocation(&averageNeighborLocation, AGENT_COHESION_RADIUS);
-    if (!neighborsWithinRange) {
-        return {0, 0};
-    }
-
-    //Create a vector between this agent and the average position of the agents within range
-    argos::CVector2 vectorToOtherAgents =
-            argos::CVector2(averageNeighborLocation.x, averageNeighborLocation.y)
-            - argos::CVector2(this->position.x, this->position.y);
-
-    return vectorToOtherAgents;
-}
-
-
-/**
- * Calculate the vector to avoid other agents:
- * If the distance between other agents is less than a certain threshold,
- * create a vector in the opposite direction of the average location of these agents.
- * @return a vector pointing away from the average location other agents
- */
-argos::CVector2 Agent::calculateAgentAvoidanceVector() {
-
-    Coordinate averageNeighborLocation = {0, 0};
-
-    bool neighborsWithinRange = getAverageNeighborLocation(&averageNeighborLocation, AGENT_AVOIDANCE_RADIUS);
-    if (!neighborsWithinRange) {
-        return {0, 0};
-    }
-
-    //Create a vector between this agent and the average position of the agents within range
-    argos::CVector2 vectorToOtherAgents =
-            argos::CVector2(averageNeighborLocation.x, averageNeighborLocation.y)
-            - argos::CVector2(this->position.x, this->position.y);
-
-    return vectorToOtherAgents * -1;
-}
-
-/**
- * Calculate the vector to align with other agents:
- * If the distance between other agents is less than a certain threshold,
- * create a vector in the direction of the average vector of these agents, considering speed.
- * @return
- */
-argos::CVector2 Agent::calculateAgentAlignmentVector() {
-    argos::CVector2 alignmentVector = {0, 0};
-    int nAgentsWithinRange = 0;
-
-
-    //Get the velocities of the agents within range
-    for (const auto &agentVelocity: agentVelocities) {
-        std::string agentID = agentVelocity.first;
-        Coordinate otherAgentLocation = agentLocations[agentID].first;
-        argos::CVector2 agentVector = agentVelocity.second.first;
-        double agentSpeed = agentVelocity.second.second;
-        argos::CVector2 vectorToOtherAgent = argos::CVector2(otherAgentLocation.x, otherAgentLocation.y)
-                                             - argos::CVector2(this->position.x, this->position.y);
-        if (vectorToOtherAgent.Length() < AGENT_ALIGNMENT_RADIUS) {
-            alignmentVector += agentVector * agentSpeed;
-            nAgentsWithinRange++;
-        }
-    }
-
-    //If no agents are within range, there is no average velocity
-    if (nAgentsWithinRange == 0) {
-        return {0, 0};
-    }
-
-    //Else calculate the average velocity of the agents within range
-    alignmentVector /= nAgentsWithinRange;
-    return alignmentVector;
-}
-
-// Union-Find (Disjoint-Set) data structure
-class UnionFind {
-public:
-    void add(int x) {
-        if (parent.find(x) == parent.end()) {
-            parent[x] = x;
-            rank[x] = 0;
-        }
-    }
-
-    int find(int x) {
-        if (parent[x] != x) {
-            parent[x] = find(parent[x]);
-        }
-        return parent[x];
-    }
-
-    void unionSets(int x, int y) {
-        int rootX = find(x);
-        int rootY = find(y);
-        if (rootX != rootY) {
-            if (rank[rootX] > rank[rootY]) {
-                parent[rootY] = rootX;
-            } else if (rank[rootX] < rank[rootY]) {
-                parent[rootX] = rootY;
-            } else {
-                parent[rootY] = rootX;
-                rank[rootX]++;
-            }
-        }
-    }
-
-private:
-    std::unordered_map<int, int> parent;
-    std::unordered_map<int, int> rank;
-};
-
-void mergeAdjacentFrontiers(const std::vector<quadtree::Box> &frontiers,
-                            std::vector<std::vector<quadtree::Box>> &frontierRegions) {
-    UnionFind uf;
-    std::unordered_map<int, quadtree::Box> boxMap;
-
-    // Initialize union-find structure
-    for (int i = 0; i < frontiers.size(); ++i) {
-        uf.add(i);
-        boxMap[i] = frontiers[i];
-    }
-
-    // Check adjacency and union sets
-    for (int i = 0; i < frontiers.size(); ++i) {
-        for (int j = i + 1; j < frontiers.size(); ++j) {
-            Coordinate centerI = frontiers[i].getCenter();
-            Coordinate centerJ = frontiers[j].getCenter();
-            double distance = sqrt(pow(centerI.x - centerJ.x, 2) + pow(centerI.y - centerJ.y, 2));
-            double threshold = sqrt(2 * pow(frontiers[i].getSize() * 0.5 + frontiers[j].getSize() * 0.5, 2));
-            if (distance <= threshold) {
-                uf.unionSets(i, j);
-            }
-        }
-    }
-
-    // Group boxes by their root set representative
-    std::unordered_map<int, std::vector<quadtree::Box>> regions;
-    for (int i = 0; i < frontiers.size(); ++i) {
-        int root = uf.find(i);
-        regions[root].push_back(frontiers[i]);
-    }
-
-    // Convert to the required format
-    for (const auto &region: regions) {
-        frontierRegions.push_back(region.second);
-    }
-}
-
-argos::CVector2 Agent::calculateUnexploredFrontierVector() {
-    //According to Dynamic frontier-led swarming:
-    //https://ieeexplore-ieee-org.tudelft.idm.oclc.org/stamp/stamp.jsp?tp=&arnumber=10057179&tag=1
-    //F* = arg min (Ψ_D||p-G^f||_2 - Ψ_S * J^f) for all frontiers F^f in F
-    //F^f is a frontier, a segment that separates explored cells from unexplored cells.
-
-    //Where Ψ_D is the frontier distance weight
-    //p is the agent position
-    //G^f is the frontier position defined as G^f = (Sum(F_j^f))/J^f So the sum of the cell locations divided by the amount of cells
-    //Ψ_S is the frontier size weight
-    //J^f is the number of cells in the frontier F^f
-
-    //A cell is a frontier iff:
-    //1. Occupancy = explored
-    //2. At least one neighbor is unexplored using the 8-connected Moore neighbours. (https://en.wikipedia.org/wiki/Moore_neighborhood)
-
-    //TODO: Need to keep search area small for computation times. Maybe when in range only low scores, expand range or search a box besides.
-    std::vector<quadtree::Box> frontiers = this->quadtree->queryFrontierBoxes(this->position, FRONTIER_SEARCH_DIAMETER,
-                                                                              this->elapsed_ticks /
-                                                                              this->ticks_per_second);
-    current_frontiers = frontiers;
-
-    // Initialize an empty vector of vectors to store frontier regions
-    std::vector<std::vector<quadtree::Box>> frontierRegions = {};
-
-    mergeAdjacentFrontiers(frontiers, frontierRegions);
-
-// Iterate over each frontier box to merge adjacent ones into regions
-    for (auto frontier: frontiers) {
-        bool added = false; // Flag to check if the current frontier has been added to a region
-
-        // Iterate over existing regions to find a suitable one for the current frontier
-        for (auto &region: frontierRegions) {
-            for (auto box: region) {
-                // Calculate the center coordinates of the current box and the frontier
-                Coordinate boxCenter = box.getCenter();
-                Coordinate frontierCenter = frontier.getCenter();
-
-                // Check if the distance between the box and the frontier is less than or equal to
-                // the average diagonal of the two boxes (ensuring adjacency)
-                if (sqrt(pow(boxCenter.x - frontierCenter.x, 2) + pow(boxCenter.y - frontierCenter.y, 2)) <=
-                    sqrt(2 * pow(frontier.getSize() * 0.5 + box.getSize() * 0.5, 2))) {
-                    region.push_back(frontier); // Add the frontier to the current region
-                    added = true; // Mark the frontier as added
-                    break; // Exit the loop since the frontier has been added to a region
-                }
-            }
-        }
-
-        // If the frontier was not added to any existing region, create a new region with it
-        if (!added) {
-            frontierRegions.push_back({frontier});
-        }
-    }
-
-    current_frontier_regions = frontierRegions;
-
-    //Now we have all frontier cells merged into frontier regions
-    //Find F* by using the formula above
-    //Ψ_D = FRONTIER_DISTANCE_WEIGHT
-    //Ψ_S = FRONTIER_SIZE_WEIGHT
-
-    //Initialize variables to store the best frontier region and its score
-    std::vector<quadtree::Box> bestFrontierRegion = {};
-    Coordinate bestFrontierRegionCenter = {MAXFLOAT, MAXFLOAT};
-    double bestFrontierScore = std::numeric_limits<double>::max();
-
-    //Iterate over all frontier regions to find the best one
-    for (const auto &region: frontierRegions) {
-        //Calculate the average position of the frontier region
-        double sumX = 0;
-        double sumY = 0;
-        double totalNumberOfCellsInRegion = 0;
-        for (auto box: region) {
-            double cellsInBox = box.getSize() / quadtree->getSmallestBoxSize();
-            assert(cellsInBox == 1);
-            sumX += box.getCenter().x *
-                    cellsInBox; //Take the box size into account (parent nodes will contain the info about all its children)
-            sumY += box.getCenter().y * cellsInBox;
-            totalNumberOfCellsInRegion += cellsInBox;
-        }
-        double frontierRegionX = sumX / totalNumberOfCellsInRegion;
-        double frontierRegionY = sumY / totalNumberOfCellsInRegion;
-
-#ifdef AVOID_UNREACHABLE_FRONTIERS
-        if (frontierEvaluator.skipFrontier(this, frontierRegionX, frontierRegionY)) continue; //Skip this frontier
-#endif
-
-        //Calculate the distance between the agent and the frontier region
-        double distance = sqrt(pow(frontierRegionX - this->position.x, 2) + pow(frontierRegionY - this->position.y, 2));
-
-        //Calculate the score of the frontier region
-        double score = FRONTIER_DISTANCE_WEIGHT * distance - FRONTIER_SIZE_WEIGHT * totalNumberOfCellsInRegion;
-
-#ifdef SEPARATE_FRONTIERS
-        std::vector<double> distancesFromOtherAgents = {};
-
-        for (const auto &agentLocationPair: this->agentLocations) {
-            //Get the distance between the frontier and the last known location of other agents
-            Coordinate agentLocation = agentLocationPair.second.first;
-            double distanceFromOtherAgent = sqrt(
-                    pow(frontierRegionX - agentLocation.x, 2) + pow(frontierRegionY - agentLocation.y, 2));
-            distancesFromOtherAgents.push_back(distanceFromOtherAgent);
-        }
-
-        //If that frontier is best to visit for a different agent, skip it.
-        bool otherAgentLowerScore = false;
-        for (auto distanceFromOtherAgent: distancesFromOtherAgents) {
-            double otherAgentScore = FRONTIER_DISTANCE_WEIGHT * distanceFromOtherAgent -
-                                     FRONTIER_SIZE_WEIGHT * totalNumberOfCellsInRegion;
-            if (otherAgentScore < score) {
-                otherAgentLowerScore = true;
-                break;
-            }
-        }
-        if (otherAgentLowerScore) continue;
-#endif
-
-        //If the score is lower than the best score, update the best score and best frontier region
-        if (score < bestFrontierScore) {
-            bestFrontierScore = score;
-            bestFrontierRegionCenter = {frontierRegionX, frontierRegionY};
-        }
-    }
-
-
-    this->currentBestFrontier = bestFrontierRegionCenter;
-
-
-
-    //Calculate the vector to the best frontier region
-    argos::CVector2 vectorToBestFrontier = argos::CVector2(bestFrontierRegionCenter.x, bestFrontierRegionCenter.y)
-                                           - argos::CVector2(this->position.x, this->position.y);
-
-    return vectorToBestFrontier;
-}
 
 bool Agent::frontierPheromoneEvaporated() {
     quadtree->queryFrontierBoxes(this->currentBestFrontier, quadtree->getSmallestBoxSize(),
@@ -903,10 +430,10 @@ void Agent::calculateNextPosition() {
     //4. Repulsion from objects/walls (done)
 
 
-    argos::CVector2 virtualWallAvoidanceVector = getVirtualWallAvoidanceVector();
-    argos::CVector2 agentCohesionVector = calculateAgentCohesionVector();
-    argos::CVector2 agentAvoidanceVector = calculateAgentAvoidanceVector();
-    argos::CVector2 agentAlignmentVector = calculateAgentAlignmentVector();
+    argos::CVector2 virtualWallAvoidanceVector = ForceVectorCalculator::getVirtualWallAvoidanceVector(this);
+    argos::CVector2 agentCohesionVector = ForceVectorCalculator::calculateAgentCohesionVector(this);
+    argos::CVector2 agentAvoidanceVector = ForceVectorCalculator::calculateAgentAvoidanceVector(this);
+    argos::CVector2 agentAlignmentVector = ForceVectorCalculator::calculateAgentAlignmentVector(this);
 
     argos::CVector2 unexploredFrontierVector = argos::CVector2(this->currentBestFrontier.x - this->position.x,
                                                                this->currentBestFrontier.y - this->position.y);
@@ -937,11 +464,11 @@ frontierEvaluator.frontierHasLowConfidenceOrAvoiding(this) ||
         //If we are not currently wall following
         if (wallFollower.wallFollowingDirection == 0) {
             //Find new frontier
-            unexploredFrontierVector = calculateUnexploredFrontierVector();
+            unexploredFrontierVector = ForceVectorCalculator::calculateUnexploredFrontierVector(this);
         }
 #else
         //Find new frontier
-        unexploredFrontierVector = calculateUnexploredFrontierVector();
+        unexploredFrontierVector = ForceVectorCalculator::calculateUnexploredFrontierVector();
 #endif
     }
 
@@ -950,11 +477,11 @@ frontierEvaluator.frontierHasLowConfidenceOrAvoiding(this) ||
 #ifdef WALL_FOLLOWING_ENABLED
     if (this->currentBestFrontier == Coordinate{MAXFLOAT, MAXFLOAT} || wallFollower.wallFollowingDirection == 0) {
         //Find new frontier
-        unexploredFrontierVector = calculateUnexploredFrontierVector();
+        unexploredFrontierVector = ForceVectorCalculator::calculateUnexploredFrontierVector();
     }
 #else
     //Find new frontier
-        unexploredFrontierVector = calculateUnexploredFrontierVector();
+        unexploredFrontierVector = ForceVectorCalculator::calculateUnexploredFrontierVector();
 #endif
 #endif
 #ifdef WALKING_STATE_WHEN_NO_FRONTIERS
@@ -987,7 +514,7 @@ frontierEvaluator.frontierHasLowConfidenceOrAvoiding(this) ||
                                                         agentAlignmentVector,
                                                         unexploredFrontierVector);
     argos::CRadians objectAvoidanceAngle;
-    bool isThereAFreeAngle = calculateObjectAvoidanceAngle(&objectAvoidanceAngle, total_vector.Angle());
+    bool isThereAFreeAngle = ForceVectorCalculator::calculateObjectAvoidanceAngle(this, &objectAvoidanceAngle, total_vector.Angle());
 
     //If there is not a free angle to move to, do not move
     if (!isThereAFreeAngle) {
